+++ server/models/collection.js
@@ -1,0 +1,31 @@
+//
+// LibreTexts Conductor
+// collection.js
+// Mongoose Model
+//
+
+const mongoose = require('mongoose');
+const Schema = mongoose.Schema;
+
+const CollectionSchema = new Schema({
+    orgID: {                        // the organization's internal identifier string
+        type: String,
+        required: true
+    },
+    colID: {                        // base62 8-digit identifier
+        type: String,
+        required: true
+    },
+    title: {                        // the collection title/name
+        type: String,
+        required: true
+    },
+    coverPhoto: String,             // the collection's "cover photo"/thumbnail
+    resources: [String]             // the array of resource IDs included in the collection
+}, {
+    timestamps: true
+});
+
+const Collection = mongoose.model('Collection', CollectionSchema);
+
+module.exports = Collection;